Deprecate all base classes for the service test suite.
First step in switching over to the ApplicationBuilder.
