+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/ApplicationServiceCRUDTest.java
@@ -43,7 +43,7 @@ import java.util.UUID;
  */
 @RunWith(Parameterized.class)
 public final class ApplicationServiceCRUDTest
-        extends AbstractServiceCRUDTest<Application> {
+        extends DAbstractServiceCRUDTest<Application> {
 
     /**
      * Create a new instance of this parameterized test.
