+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/ClientServiceCRUDTest.java
@@ -41,7 +41,7 @@ import java.util.UUID;
  * @author Michael Krotscheck
  */
 public final class ClientServiceCRUDTest
-        extends AbstractServiceCRUDTest<Client> {
+        extends DAbstractServiceCRUDTest<Client> {
 
     /**
      * Create a new instance of this parameterized test.
