+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/DAbstractResourceTest.java
@@ -67,7 +67,8 @@ import java.util.stream.Collectors;
  *
  * @author Michael Krotscheck
  */
-public abstract class AbstractResourceTest extends DContainerTest {
+@Deprecated
+public abstract class DAbstractResourceTest extends DContainerTest {
 
     /**
      * The Admin context.
