+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/DAbstractServiceBrowseTest.java
@@ -18,20 +18,22 @@
 
 package net.krotscheck.kangaroo.servlet.admin.v1.resource;
 
+import net.krotscheck.kangaroo.common.response.ApiParam;
 import net.krotscheck.kangaroo.database.entity.AbstractEntity;
+import net.krotscheck.kangaroo.database.entity.Client;
 import net.krotscheck.kangaroo.database.entity.ClientType;
 import net.krotscheck.kangaroo.database.entity.OAuthToken;
 import net.krotscheck.kangaroo.database.entity.User;
 import net.krotscheck.kangaroo.servlet.admin.v1.Scope;
 import net.krotscheck.kangaroo.test.EnvironmentBuilder;
-import org.apache.lucene.search.Query;
-import org.hibernate.search.query.dsl.QueryBuilder;
+import net.krotscheck.kangaroo.test.HttpUtil;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 import javax.ws.rs.core.GenericType;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import java.util.ArrayList;
@@ -39,7 +41,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
 /**
@@ -48,14 +49,10 @@ import java.util.stream.Collectors;
  * @param <T> The type of entity to execute this test for.
  * @author Michael Krotscheck
  */
+@Deprecated
 @RunWith(Parameterized.class)
-public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
-        extends AbstractResourceTest {
-
-    /**
-     * Class reference for this class' type, used in casting.
-     */
-    private final Class<T> typingClass;
+public abstract class DAbstractServiceBrowseTest<T extends AbstractEntity>
+        extends DAbstractResourceTest {
 
     /**
      * The scope to grant the issued token.
@@ -66,6 +63,11 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
      * The type of OAuth2 client to create.
      */
     private final ClientType clientType;
+
+    /**
+     * The client under test.
+     */
+    private Client client;
 
     /**
      * Whether to create a user, or fall back on an existing user. In most
@@ -86,16 +88,13 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
     /**
      * Create a new instance of this parameterized test.
      *
-     * @param typingClass The raw class type, used for type-based parsing.
-     * @param clientType  The type of  client.
-     * @param tokenScope  The client scope to issue.
-     * @param createUser  Whether to create a new user.
+     * @param clientType The type of  client.
+     * @param tokenScope The client scope to issue.
+     * @param createUser Whether to create a new user.
      */
-    public AbstractServiceSearchTest(final Class<T> typingClass,
-                                     final ClientType clientType,
-                                     final String tokenScope,
-                                     final Boolean createUser) {
-        this.typingClass = typingClass;
+    public DAbstractServiceBrowseTest(final ClientType clientType,
+                                      final String tokenScope,
+                                      final Boolean createUser) {
         this.tokenScope = tokenScope;
         this.clientType = clientType;
         this.createUser = createUser;
@@ -107,6 +106,23 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
      * @return The list type, used for test decoding.
      */
     protected abstract GenericType<List<T>> getListType();
+
+    /**
+     * Return the list of entities which should be accessible given a
+     * specific token.
+     *
+     * @param token The oauth token to test against.
+     * @return A list of entities (could be empty).
+     */
+    protected abstract List<T> getAccessibleEntities(OAuthToken token);
+
+    /**
+     * Return the list of entities which are owned by the given oauth token.
+     *
+     * @param owner The owner of these entities.
+     * @return A list of entities (could be empty).
+     */
+    protected abstract List<T> getOwnedEntities(User owner);
 
     /**
      * Return the list of entities which should be accessible given a
@@ -146,56 +162,6 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
     }
 
     /**
-     * Return the list of entities which should be accessible given a
-     * specific token.
-     *
-     * @param token The oauth token to test against.
-     * @return A list of entities (could be empty).
-     */
-    protected abstract List<T> getAccessibleEntities(OAuthToken token);
-
-    /**
-     * Return the list of entities which are owned by the given oauth token.
-     *
-     * @param owner The owner of these entities.
-     * @return A list of entities (could be empty).
-     */
-    protected abstract List<T> getOwnedEntities(User owner);
-
-    /**
-     * Return the list of field names on which this particular entity type
-     * has build a search index.
-     *
-     * @return An array of field names.
-     */
-    protected abstract String[] getSearchIndexFields();
-
-    /**
-     * Run a raw search query against our lucene index, given a specific string.
-     *
-     * @param query The search query to apply.
-     * @return True if it's a member of a result set, otherwise false.
-     */
-    protected final List<T> getSearchResults(final String query) {
-        QueryBuilder b = getSearchFactory()
-                .buildQueryBuilder()
-                .forEntity(typingClass)
-                .get();
-
-        Query luceneQuery = b
-                .keyword()
-                .fuzzy()
-                .onFields(getSearchIndexFields())
-                .matching(query)
-                .createQuery();
-
-        List<T> results = (List<T>) getFullTextSession()
-                .createFullTextQuery(luceneQuery, typingClass)
-                .list();
-        return results;
-    }
-
-    /**
      * Return the second application context (not the admin context).
      *
      * @return The secondary context in this test.
@@ -220,6 +186,24 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
      */
     protected final OAuthToken getSecondaryToken() {
         return otherApp.getToken();
+    }
+
+    /**
+     * Return the client under test.
+     *
+     * @return The client under test.
+     */
+    public final Client getAdminClient() {
+        return client;
+    }
+
+    /**
+     * Retrieve the scope under test.
+     *
+     * @return The scope that's being tested.
+     */
+    public final String getTokenScope() {
+        return tokenScope;
     }
 
     /**
@@ -250,11 +234,36 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
 
         // Get the admin app and create users based on the configured
         // parameters.
-        EnvironmentBuilder context = getAdminContext().client(clientType);
+        EnvironmentBuilder context = getAdminContext()
+                .client(clientType);
+
+        client = context.getClient();
+
+        // Create a whole lot of applications to run some tests against.
+        for (int i = 0; i < 10; i++) {
+            String appName = String.format("Application %s- %s", i, i % 2 == 0
+                    ? "many" : "frown");
+            fixtures.add(new EnvironmentBuilder(getSession(), appName)
+                    .owner(adminApp.getUser()));
+        }
+        fixtures.add(new EnvironmentBuilder(getSession(), "Single")
+                .owner(adminApp.getUser()));
+
+
         if (createUser) {
             context.user().identity();
         }
         adminAppToken = context.bearerToken(tokenScope).getToken();
+
+        // Create a whole lot more applications to run some tests against.
+        for (int i = 0; i < 10; i++) {
+            String appName = String.format("Application %s- %s", i, i % 2 == 0
+                    ? "many" : "frown");
+            fixtures.add(new EnvironmentBuilder(getSession(), appName)
+                    .owner(adminApp.getUser()));
+        }
+        fixtures.add(new EnvironmentBuilder(getSession(), "Single")
+                .owner(adminApp.getUser()));
 
         // Create a second app, owned by another user.
         otherApp = new EnvironmentBuilder(getSession())
@@ -298,20 +307,6 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
         otherApp.role("Second Role - many");
         otherApp.role("Third Role - many");
         otherApp.role("Fourth Role - many");
-
-        // Create a whole lot of applications to run some tests against.
-        for (int i = 0; i < 10; i++) {
-            String appName = String.format("Application %s- %s", i, i % 2 == 0
-                    ? "many" : "frown");
-            fixtures.add(new EnvironmentBuilder(getSession(), appName)
-                    .owner(adminApp.getOwner()));
-            fixtures.add(new EnvironmentBuilder(getSession(), appName)
-                    .owner(adminApp.getUser()));
-        }
-        fixtures.add(new EnvironmentBuilder(getSession(), "Single")
-                .owner(adminApp.getOwner()));
-        fixtures.add(new EnvironmentBuilder(getSession(), "Single")
-                .owner(adminApp.getUser()));
 
         // Create some users
         context.user()
@@ -358,309 +353,284 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
                 .authToken();
 
         fixtures.add(otherApp);
-
         return fixtures;
     }
 
     /**
-     * Test that searching with no query throws an error.
+     * Assert that you can browse applications.
      */
     @Test
-    public final void testSearchNoQuery() {
+    public final void testBrowse() {
         Map<String, String> params = new HashMap<>();
-
-        Response r = search(params, getAdminToken());
-        assertErrorResponse(r, Status.BAD_REQUEST);
-    }
-
-    /**
-     * Test that searching with a stopword returns an error.
-     */
-    @Test
-    public final void testSearchStopWord() {
-        Map<String, String> params = new HashMap<>();
-        params.put("q", "with");
-
-        Response r = search(params, getAdminToken());
-        assertErrorResponse(r, Status.BAD_REQUEST);
-    }
-
-    /**
-     * Test a simple search.
-     *
-     * @throws Exception Unexpected exception.
-     */
-    @Test
-    public final void testSearch() throws Exception {
-        String query = "many";
-        OAuthToken token = getAdminToken();
-        Map<String, String> params = new HashMap<>();
-        params.put("q", query);
-        Response r = search(params, token);
-
-        // Determine result set.
-        List<T> searchResults = getSearchResults(query);
-        List<T> accessibleEntities = getAccessibleEntities(token);
-
-        List<T> expectedResults = searchResults
-                .stream()
-                .filter((item) -> accessibleEntities.indexOf(item) > -1)
-                .collect(Collectors.toList());
-
-        Integer expectedTotal = expectedResults.size();
-        int expectedResultSize = Math.min(10, expectedTotal);
-        Integer expectedOffset = 0;
-        Integer expectedLimit = 10;
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
 
         if (isLimitedByClientCredentials()) {
             assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
                     "invalid_scope");
         } else {
-            Assert.assertTrue(expectedTotal > 1);
-
             List<T> results = r.readEntity(getListType());
-            Assert.assertEquals(expectedOffset.toString(),
-                    r.getHeaderString("Offset"));
-            Assert.assertEquals(expectedLimit.toString(),
-                    r.getHeaderString("Limit"));
-            Assert.assertEquals(expectedTotal.toString(),
-                    r.getHeaderString("Total"));
-            Assert.assertEquals(expectedResultSize, results.size());
-        }
-    }
-
-    /**
-     * Test that we can offset our search results.
-     */
-    @Test
-    public final void testSearchOffset() {
-        String query = "many";
-        Integer offset = 1;
-        OAuthToken token = getAdminToken();
-        Map<String, String> params = new HashMap<>();
-        params.put("q", query);
-        params.put("offset", offset.toString());
-        Response r = search(params, token);
-
-        // Determine result set.
-        List<T> searchResults = getSearchResults(query);
-        List<T> accessibleEntities = getAccessibleEntities(token);
-
-        List<T> expectedResults = searchResults
-                .stream()
-                .filter((item) -> accessibleEntities.indexOf(item) > -1)
-                .collect(Collectors.toList());
-
-        Integer expectedTotal = expectedResults.size();
-        int expectedResultSize = Math.min(10, expectedTotal - offset);
-        Integer expectedOffset = offset;
-        Integer expectedLimit = 10;
-
-        if (isLimitedByClientCredentials()) {
-            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
-                    "invalid_scope");
-        } else {
-            Assert.assertTrue(expectedTotal > 1);
-
-            List<T> results = r.readEntity(getListType());
-            Assert.assertEquals(expectedOffset.toString(),
-                    r.getHeaderString("Offset"));
-            Assert.assertEquals(expectedLimit.toString(),
-                    r.getHeaderString("Limit"));
-            Assert.assertEquals(expectedTotal.toString(),
-                    r.getHeaderString("Total"));
-            Assert.assertEquals(expectedResultSize, results.size());
-        }
-    }
-
-    /**
-     * Test that we can limit our search results.
-     */
-    @Test
-    public final void testSearchLimit() {
-        String query = "many";
-        Integer limit = 1;
-        OAuthToken token = getAdminToken();
-        Map<String, String> params = new HashMap<>();
-        params.put("q", query);
-        params.put("limit", limit.toString());
-        Response r = search(params, token);
-
-        // Determine result set.
-        List<T> searchResults = getSearchResults(query);
-        List<T> accessibleEntities = getAccessibleEntities(token);
-
-        List<T> expectedResults = searchResults
-                .stream()
-                .filter((item) -> accessibleEntities.indexOf(item) > -1)
-                .collect(Collectors.toList());
-
-        Integer expectedTotal = expectedResults.size();
-        int expectedResultSize = limit;
-        Integer expectedOffset = 0;
-        Integer expectedLimit = limit;
-
-        if (isLimitedByClientCredentials()) {
-            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
-                    "invalid_scope");
-        } else {
-            Assert.assertTrue(expectedTotal > 1);
-
-            List<T> results = r.readEntity(getListType());
-            Assert.assertEquals(expectedOffset.toString(),
-                    r.getHeaderString("Offset"));
-            Assert.assertEquals(expectedLimit.toString(),
-                    r.getHeaderString("Limit"));
-            Assert.assertEquals(expectedTotal.toString(),
-                    r.getHeaderString("Total"));
-            Assert.assertEquals(expectedResultSize, results.size());
-        }
-    }
-
-    /**
-     * Test that only an admin scope can search by a different user.
-     */
-    @Test
-    public final void testSearchByOwner() {
-        String query = "many";
-        OAuthToken adminToken = getAdminToken();
-        User secondaryOwner = getSecondaryContext().getOwner();
-        Map<String, String> params = new HashMap<>();
-        params.put("q", query);
-        params.put("owner", getAdminContext().getOwner().getId().toString());
-
-        // Determine result set.
-        List<T> searchResults = getSearchResults(query);
-        List<T> ownedEntities = getOwnedEntities(secondaryOwner);
-        List<T> accessibleEntities = getAccessibleEntities(adminToken);
-
-        List<T> expectedResults = searchResults
-                .stream()
-                .filter((item) -> ownedEntities.indexOf(item) > -1)
-                .filter((item) -> accessibleEntities.indexOf(item) > -1)
-                .collect(Collectors.toList());
-
-        Integer expectedTotal = expectedResults.size();
-        int expectedResultSize = Math.min(10, expectedTotal);
-        Integer expectedOffset = 0;
-        Integer expectedLimit = 10;
-
-        // Make request.
-        Response r = search(params, adminToken);
-
-        if (isLimitedByClientCredentials()) {
-            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
-                    "invalid_scope");
-        } else if (!isAccessible(secondaryOwner, adminToken)) {
-            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
-                    "invalid_scope");
-        } else {
-            Assert.assertTrue(expectedTotal > 1);
-
-            List<T> results = r.readEntity(getListType());
-            Assert.assertEquals(200, r.getStatus());
-            Assert.assertEquals(expectedTotal.toString(),
-                    r.getHeaderString("Total"));
-            Assert.assertEquals(expectedOffset.toString(),
-                    r.getHeaderString("Offset"));
-            Assert.assertEquals(expectedLimit.toString(),
-                    r.getHeaderString("Limit"));
-            Assert.assertEquals(expectedResultSize, results.size());
-        }
-    }
-
-    /**
-     * Test that any scope can search by their own owner ID.
-     */
-    @Test
-    public final void testSearchBySelf() {
-        String query = "single";
-        OAuthToken adminToken = getAdminToken();
-        if (adminToken.getIdentity() != null) {
-            Map<String, String> params = new HashMap<>();
-            params.put("q", query);
-            params.put("owner", adminToken.getIdentity()
-                    .getUser().getId().toString());
-
-            // Determine result set.s
-            List<T> searchResults = getSearchResults(query);
-            List<T> ownedEntities = getOwnedEntities(adminToken);
-            List<T> expectedResults = searchResults
-                    .stream()
-                    .filter((item) -> ownedEntities.indexOf(item) > -1)
-                    .collect(Collectors.toList());
-            Integer expectedCount = expectedResults.size();
-
-            // Make request.
-            Response r = search(params, adminToken);
-
-            Assert.assertTrue(expectedCount > 0);
-
-            List<T> results = r.readEntity(getListType());
-            Assert.assertEquals(200, r.getStatus());
-            Assert.assertEquals(expectedCount.toString(),
-                    r.getHeaderString("Total"));
             Assert.assertEquals("0", r.getHeaderString("Offset"));
             Assert.assertEquals("10", r.getHeaderString("Limit"));
-            Assert.assertEquals(Math.min(10, expectedCount), results.size());
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults, 10), results.size());
+        }
+    }
+
+    /**
+     * Assert that you can browse with a limit set.
+     */
+    @Test
+    public final void testBrowseLimit() {
+        Integer limit = 2;
+        Map<String, String> params = new HashMap<>();
+        params.put("limit", limit.toString());
+
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
+
+        if (isLimitedByClientCredentials()) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals("0", r.getHeaderString("Offset"));
+            Assert.assertEquals(limit.toString(), r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults, limit),
+                    results.size());
+        }
+    }
+
+    /**
+     * Assert that you can browse with an offset.
+     */
+    @Test
+    public final void testBrowseOffset() {
+        Integer offset = 1;
+        Map<String, String> params = new HashMap<>();
+        params.put("offset", offset.toString());
+
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
+
+        if (isLimitedByClientCredentials()) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals(offset.toString(),
+                    r.getHeaderString("Offset"));
+            Assert.assertEquals("10", r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults - offset, 10),
+                    results.size());
+        }
+    }
+
+    /**
+     * Assert that you can browse and sort default ascending.
+     */
+    @Test
+    public final void testBrowseSortDefault() {
+        Map<String, String> params = new HashMap<>();
+        params.put(ApiParam.SORT_QUERY, "createdDate");
+
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
+
+        if (isLimitedByClientCredentials()) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals("0", r.getHeaderString("Offset"));
+            Assert.assertEquals("10", r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults, 10),
+                    results.size());
+
+            results.stream().sorted((e1, e2) -> e1.getCreatedDate()
+                    .compareTo(e2.getCreatedDate()));
+        }
+    }
+
+    /**
+     * Assert that you can browse and sort ascending.
+     */
+    @Test
+    public final void testBrowseSortAscending() {
+        Map<String, String> params = new HashMap<>();
+        params.put(ApiParam.SORT_QUERY, "createdDate");
+        params.put(ApiParam.ORDER_QUERY, "ASC");
+
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
+
+        if (isLimitedByClientCredentials()) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals("0", r.getHeaderString("Offset"));
+            Assert.assertEquals("10", r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults, 10),
+                    results.size());
+
+            results.stream().sorted((e1, e2) -> e1.getCreatedDate()
+                    .compareTo(e2.getCreatedDate()));
+        }
+    }
+
+    /**
+     * Assert that you can browse and sort descending.
+     */
+    @Test
+    public final void testBrowseSortDescending() {
+        Map<String, String> params = new HashMap<>();
+        params.put(ApiParam.SORT_QUERY, "createdDate");
+        params.put(ApiParam.ORDER_QUERY, "DESC");
+
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
+
+        if (clientType.equals(ClientType.ClientCredentials)
+                && tokenScope.equals(getRegularScope())) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals("0", r.getHeaderString("Offset"));
+            Assert.assertEquals("10", r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults, 10),
+                    results.size());
+
+            results.stream().sorted((e1, e2) -> e2.getCreatedDate()
+                    .compareTo(e1.getCreatedDate()));
+        }
+    }
+
+    /**
+     * Assert that an invalid sort parameter defaults to ASC.
+     */
+    @Test
+    public final void testBrowseSortInvalid() {
+        Map<String, String> params = new HashMap<>();
+        params.put(ApiParam.SORT_QUERY, "createdDate");
+        params.put(ApiParam.ORDER_QUERY, "invalid_sort");
+
+        Response r = browse(params, adminAppToken);
+        Integer expectedResults = getAccessibleEntities(adminAppToken)
+                .size();
+
+        if (isLimitedByClientCredentials()) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals("0", r.getHeaderString("Offset"));
+            Assert.assertEquals("10", r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedResults.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedResults, 10),
+                    results.size());
+
+            results.stream().sorted((e1, e2) -> e1.getCreatedDate()
+                    .compareTo(e2.getCreatedDate()));
+        }
+    }
+
+    /**
+     * Ensure that we can filter by an entity's owner.
+     */
+    @Test
+    public final void testBrowseFilterByOwner() {
+        // Sometimes we own this application, sometimes we don't.
+        User owner = getAdminContext().getApplication().getOwner();
+
+        Map<String, String> params = new HashMap<>();
+        params.put("owner", owner.getId().toString());
+        Response r = browse(params, getAdminToken());
+
+        List<T> accessibleEntities = getAccessibleEntities(getAdminToken());
+        List<T> expectedEntities = accessibleEntities
+                .stream()
+                .filter(e -> owner.equals(e.getOwner()))
+                .collect(Collectors.toList());
+        Integer expectedEntityCount = expectedEntities.size();
+
+        if (isLimitedByClientCredentials()
+                || !isAccessible(owner, getAdminToken())) {
+            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
+                    "invalid_scope");
+        } else {
+            Assert.assertTrue(expectedEntityCount > 0);
+
+            List<T> results = r.readEntity(getListType());
+            Assert.assertEquals("0", r.getHeaderString("Offset"));
+            Assert.assertEquals("10", r.getHeaderString("Limit"));
+            Assert.assertEquals(expectedEntityCount.toString(),
+                    r.getHeaderString("Total"));
+            Assert.assertEquals(Math.min(expectedEntityCount, 10),
+                    results.size());
+        }
+    }
+
+    /**
+     * Test that searches by malformed owners return appropriate responses.
+     */
+    @Test
+    public final void testBrowseFilterByMalformedOwner() {
+        Response response = target(getUrlForId(null))
+                .queryParam("owner", "malformed")
+                .request()
+                .header(HttpHeaders.AUTHORIZATION,
+                        HttpUtil.authHeaderBearer(adminAppToken.getId()))
+                .get();
+
+        assertErrorResponse(response, Status.NOT_FOUND);
+    }
+
+    /**
+     * Test that any scope can browse by their own ID.
+     */
+    @Test
+    public final void testBrowseBySelf() {
+        if (adminAppToken.getIdentity() != null) {
+            Response response = target(getUrlForId(null))
+                    .queryParam("owner",
+                            adminAppToken.getIdentity().getUser().getId())
+                    .request()
+                    .header(HttpHeaders.AUTHORIZATION,
+                            HttpUtil.authHeaderBearer(adminAppToken.getId()))
+                    .get();
+
+            Integer expectedResults = getOwnedEntities(adminAppToken)
+                    .size();
+
+            List<T> results = response.readEntity(getListType());
+            Assert.assertEquals(200, response.getStatus());
+            Assert.assertEquals(expectedResults.toString(),
+                    response.getHeaderString("Total"));
+            Assert.assertEquals("0", response.getHeaderString("Offset"));
+            Assert.assertEquals("10", response.getHeaderString("Limit"));
+            Assert.assertEquals(Math.min(10, expectedResults), results.size());
         } else {
             Assert.assertTrue(true);
-        }
-    }
-
-    /**
-     * Test that searches by an invalid owner.
-     */
-    // TODO(krotscheck): This should return a 400.
-    @Test
-    public final void testSearchByInvalidOwner() {
-        OAuthToken token = getAdminToken();
-        Map<String, String> params = new HashMap<>();
-        params.put("q", "many");
-        params.put("owner", UUID.randomUUID().toString());
-
-        Response r = search(params, token);
-
-        if (token.getScopes().keySet().contains(getAdminScope())) {
-            assertErrorResponse(r, Status.BAD_REQUEST);
-        } else {
-            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
-                    "invalid_scope");
-        }
-    }
-
-    /**
-     * Test that searches by malformed owners returns a 404.
-     */
-    // TODO(krotscheck): This should return a 400.
-    @Test
-    public final void testSearchByMalformedOwner() {
-        Map<String, String> params = new HashMap<>();
-        params.put("q", "single");
-        params.put("owner", "malformed");
-
-        Response r = search(params, getAdminToken());
-        assertErrorResponse(r, Status.NOT_FOUND);
-    }
-
-    /**
-     * Test that bad search results are empty.
-     */
-    @Test
-    public final void testSearchNoResults() {
-        Map<String, String> params = new HashMap<>();
-        params.put("q", "lolcat");
-        Response r = search(params, getAdminToken());
-
-        if (isLimitedByClientCredentials()) {
-            assertErrorResponse(r, Status.BAD_REQUEST.getStatusCode(),
-                    "invalid_scope");
-        } else {
-            List<T> results = r.readEntity(getListType());
-            Assert.assertEquals("0", r.getHeaderString("Offset"));
-            Assert.assertEquals("10", r.getHeaderString("Limit"));
-            Assert.assertEquals("0", r.getHeaderString("Total"));
-            Assert.assertEquals(0, results.size());
         }
     }
 
@@ -668,11 +638,11 @@ public abstract class AbstractServiceSearchTest<T extends AbstractEntity>
      * Assert that we can only browse the resource if we're logged in.
      */
     @Test
-    public final void testSearchNoAuth() {
-        Map<String, String> params = new HashMap<>();
-        params.put("q", "single");
+    public final void testBrowseNoAuth() {
+        Response response = target(getUrlForId(null))
+                .request()
+                .get();
 
-        Response r = search(params, null);
-        assertErrorResponse(r, Status.FORBIDDEN);
+        assertErrorResponse(response, Status.FORBIDDEN);
     }
 }
