+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/OAuthTokenServiceBrowseTest.java
@@ -48,7 +48,7 @@ import java.util.stream.Collectors;
  */
 @RunWith(Parameterized.class)
 public final class OAuthTokenServiceBrowseTest
-        extends AbstractServiceBrowseTest<OAuthToken> {
+        extends DAbstractServiceBrowseTest<OAuthToken> {
 
     /**
      * Generic type declaration for list decoding.
