+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/OAuthTokenServiceSearchTest.java
@@ -50,7 +50,7 @@ import java.util.stream.Collectors;
  */
 @RunWith(Parameterized.class)
 public final class OAuthTokenServiceSearchTest
-        extends AbstractServiceSearchTest<OAuthToken> {
+        extends DAbstractServiceSearchTest<OAuthToken> {
 
     /**
      * Convenience generic type for response decoding.
