+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/RoleServiceBrowseTest.java
@@ -49,7 +49,7 @@ import java.util.stream.Collectors;
  */
 @RunWith(Parameterized.class)
 public final class RoleServiceBrowseTest
-        extends AbstractServiceBrowseTest<Role> {
+        extends DAbstractServiceBrowseTest<Role> {
 
     /**
      * Generic type declaration for list decoding.
