+++ kangaroo-server-admin/src/test/java/net/krotscheck/kangaroo/servlet/admin/v1/resource/UserServiceBrowseTest.java
@@ -49,7 +49,7 @@ import java.util.stream.Collectors;
  */
 @RunWith(Parameterized.class)
 public final class UserServiceBrowseTest
-        extends AbstractServiceBrowseTest<User> {
+        extends DAbstractServiceBrowseTest<User> {
 
     /**
      * Generic type declaration for list decoding.
